+++ plugins/11_TheWarWithin/CHANGELOG.md
@@ -1,6 +1,8 @@
 # v123
 
-* Localization updates: esES, frFR
+* Update spawn points for several rares in K'aresh
+* Added additional spawn of Abandoned Lockbox and Lightly-Dented Luggage treasures in K'aresh
+* Localization updates: deDE, esES, frFR, zhCN
 
 # v122